Initial support for loading Add data to a block
When using Forge mods (before 1.13?) with modded blocks, the block ID becomes split into the Add section.

This commit adds support for reading that block ID fully, even if minecraft-data does not have that block ID stored. However, I'm not sure on how to write it into a buffer, and as a result, this commit is currently incomplete.